+++ app/Http/Controllers/IncomeController.php
@@ -9,6 +9,7 @@ use App\Http\Requests\Income\UpdateRequest;
 use App\Http\Resources\IncomeResource;
 use App\Http\Resources\IncomeResourceCollection;
 use App\Repositories\Contracts\IncomeRepository;
+use Illuminate\Auth\Access\AuthorizationException;
 
 class IncomeController extends Controller
 {
@@ -31,9 +32,12 @@ class IncomeController extends Controller
      *
      * @param IndexRequest $request
      * @return IncomeResourceCollection
+     * @throws AuthorizationException
      */
     public function index(IndexRequest $request)
     {
+        $this->authorize('list', [Income::class, $request->get('propertyId')]);
+
         $incomes = $this->incomeRepository->findBy($request->all());
 
         return new IncomeResourceCollection($incomes);
@@ -44,9 +48,12 @@ class IncomeController extends Controller
      *
      * @param  \Illuminate\Http\Request  $request
      * @return IncomeResource
+     * @throws AuthorizationException
      */
     public function store(StoreRequest $request)
     {
+        $this->authorize('store', [Income::class, $request->get('propertyId')]);
+
         $income = $this->incomeRepository->save($request->all());
 
         return new IncomeResource($income);
@@ -57,9 +64,12 @@ class IncomeController extends Controller
      *
      * @param Income $income
      * @return IncomeResource
+     * @throws AuthorizationException
      */
     public function show(Income $income)
     {
+        $this->authorize('show', $income);
+
         return new IncomeResource($income);
     }
 
@@ -69,9 +79,12 @@ class IncomeController extends Controller
      * @param UpdateRequest $request
      * @param Income $income
      * @return IncomeResource
+     * @throws AuthorizationException
      */
     public function update(UpdateRequest $request, Income $income)
     {
+        $this->authorize('update', $income);
+
         $income = $this->incomeRepository->update($income, $request->all());
 
         return new IncomeResource($income);
@@ -83,9 +96,12 @@ class IncomeController extends Controller
      *
      * @param Income $income
      * @return \Illuminate\Http\Response
+     * @throws AuthorizationException
      */
     public function destroy(Income $income)
     {
+        $this->authorize('destroy', $income);
+
         $this->incomeRepository->delete($income);
 
         return response()->json(null, 204);
